Move Container to own submodule
